[GREEN] Update building request calls service correctly

+++ API/WebModel/Requests/BuildingRequests/UpdateBuildingRequest.cs
@@ -5,5 +5,5 @@ namespace WebModel.Requests.BuildingRequests;
 public class UpdateBuildingRequest
 {
     public Guid ConstructionCompanyId { get; set; }
-    public double CommonExpenses { get; set; }
+    public int CommonExpenses { get; set; }
 }
